US1T3 Setup new file structure and implement model
